chore: rename lib folder to nvim-tree
Allows external access to the lua api and follow neovim plugins better
practices
