+++ lua/nvim-tree/config.lua
